pyproject.toml
[tool.poetry]
name = "ai-engineer-assignment-neeraj"
version = "0.1.0"
description = ""
authors = ["Neeraj S <nrj127@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.10"
types-pyyaml = "^6.0.12.20240311"
pyyaml = "^6.0.1"
openai = "^1.36.0"
python-dotenv = "^1.0.1"
pyboxen = "^1.3.0"
langchain-community = "^0.2.9"
langchain-openai = "^0.1.17"
pandas = "^2.2.2"
streamlit = "^1.36.0"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
